analyzeの切り分け
